dockerized, notifications added

+++ backend/files/views/delete.py
@@ -18,20 +18,25 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 
 from backend.files.models import Folder, File
+from backend.api.errors import CustomError
+from backend.static.error_codes import ERROR__API__PERMISSION__INSUFFICIENT
 
 
 class DeleteViewSet(APIView):
     def post(self, request):
         user = request.user
-        # TODO: rework this
-        root_folder = Folder.get_folder_from_path('files/' + request.data['path'], user.rlc)
-        entries = request.data['entries']
+        root_folder = Folder.get_folder_from_path(
+            "files/" + request.data["path"], user.rlc
+        )
+        if not root_folder.user_has_permission_write(user):
+            raise CustomError(ERROR__API__PERMISSION__INSUFFICIENT)
+        entries = request.data["entries"]
         for entry in entries:
-            if entry['type'] == 1:
+            if entry["type"] == 1:
                 # file
-                file = File.objects.get(pk=entry['id'])
+                file = File.objects.get(pk=entry["id"])
                 file.delete()
             else:
-                folder = Folder.objects.get(pk=entry['id'])
+                folder = Folder.objects.get(pk=entry["id"])
                 folder.delete()
-            return Response({'success': True})
+        return Response({"success": True})
